requirements.txt

openai==1.6.1
python-telegram-bot==13.7
